fixed issue with SHA256
